+++ src/Lykke.RabbitMqBroker/Subscriber/MessageReadStrategies/DeadLetteringConfigurationResult.cs
@@ -1,0 +1,3 @@
+namespace Lykke.RabbitMqBroker.Subscriber.MessageReadStrategies;
+
+internal sealed record DeadLetteringConfigurationResult(string ExchangeName);